requirements.txt
openai>=1.12.0
httpx>=0.25.0
pydantic>=2.5.0
python-dotenv>=1.0.0
docker>=7.0.0
